core/ipv4/dhcp: add hostname option to dhcp request on reboot
Unfortunately the DHCP option "hostname" (12) wasn't set in the "reboot"
dhcp message until now. dhcp_reboot is executed in each of the following
cases: REBIND, RENEW, BOUND, REBOOTING
If the dhcp server uses the hostname option in order to create matching
dns records (which is common), it wasn't possible to change its DNS
record after its initial dhcp request.
The only possibility to change the hostname was to trigger a reboot
(or a dhcp_stop).

The recent changes of 3b88e84f9b05e76a1ad96e6650a795df74af23aa made the
situation even worse. Since the last valid ip address obtained from the
DHCP server is now stored in non-volatile storage, it was not even
possible to change the hostname with a reboot of the device, since even
then a dhcp_reboot message (with the hostname option missing) is sent
(instead of a fresh dhcp request).

This fix is already implemented in the upstream version of lwip.

Signed-off-by: julian <dev@jneureuther.de>