Fix SLIP RX linked list append.
Append to the linked list of pbuf objects was broken so that the second
(and subsequent) packets received in a SLIP transaction would be
appended to themselves and consequently lost.

This resulted in these packets being dropped and the system suffering
extreme memory leakage.

Merges https://github.com/espressif/esp-lwip/pull/25